theme/minimal: add password reset bits
reset becomes a function.
Add reset_pass function to reset for password change.
info and error become functions.

Logic details:

When settings.username is blank, we are in login mode. Once
settings.username (and .password) has been filled, we are in password
change mode. If settings.password_n is blank, this is the first time we
are entering the password. Otherwise -- the second.