Major revamp
Take some of the work away from the themes. We require that themes
provide enter_user_pass (enter username and password) and enter_pass
(enter password) functions. When either one is done, they signal enter
or cancel. And we take it from there.

Other changes include (but are not limited to):

Preload username into PAM and dispense with the username function.
Emit error on reboot or poweroff failure.